refactor(agent-runtime): separate goal HTTP handlers
